Add a way to configure max request size

+++ refine-http/src/settings.rs
@@ -15,7 +15,7 @@ impl Settings {
 #[derive(Debug, Deserialize)]
 pub(crate) struct SettingsServer {
     pub(crate) port: u16,
-    pub(crate) max_request_size: u32,
+    pub(crate) max_request_size: u64,
     pub(crate) sol_lifetime: u64,
     pub(crate) sol_cleanup_interval: u64,
     pub(crate) standard_threads: usize,
